requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
groq>=0.11.0
ragas>=0.2.0
python-dotenv>=1.0.0
pypdf>=4.0.0
soundfile>=0.12.0
sentence-transformers>=3.0.0
datasets>=2.20.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
edge-tts>=7.2.7
